refactor: 내용 10자 제한 제거
