+++ src/features/homework/components/write/schemas/note.ts
@@ -1,8 +1,6 @@
 import { CourseTargetStudentInfo } from '@/features/dashboard/studynote/write/type';
-import {
-  extractTextFromTiptapJSON,
-  hasNonTextContent,
-} from '@/features/homework/hooks/use-homework-form-control';
+import { extractTextFromTiptapJSON } from '@/features/homework/hooks/use-homework-form-control';
+import { hasNonTextContent } from '@/shared/lib';
 import { JSONContent } from '@tiptap/react';
 import { z } from 'zod';
