+++ src/features/qna/schema/create.ts
@@ -1,3 +1,4 @@
+import { hasNonTextContent } from '@/shared/lib';
 import { JSONContent } from '@tiptap/react';
 import { z } from 'zod';
 
@@ -36,11 +37,12 @@ export const QnaACreateFormSchema = z.object({
   content: z.custom<JSONContent>().superRefine((val, ctx) => {
     const plainText = extractTextFromTiptapJSON(val);
     const length = plainText.trim().length;
+    const hasAttachment = hasNonTextContent(val);
 
-    if (length < 10) {
+    if (length < 1 && !hasAttachment) {
       ctx.addIssue({
         code: z.ZodIssueCode.custom,
-        message: '내용은 최소 10자 이상이어야 합니다.',
+        message: '내용 또는 첨부 파일을 입력해 주세요.',
       });
     }
 
@@ -57,11 +59,12 @@ export const QnAMessageFormSchema = z.object({
   content: z.custom<JSONContent>().superRefine((val, ctx) => {
     const plainText = extractTextFromTiptapJSON(val);
     const length = plainText.trim().length;
+    const hasAttachment = hasNonTextContent(val);
 
-    if (length < 10) {
+    if (length < 1 && !hasAttachment) {
       ctx.addIssue({
         code: z.ZodIssueCode.custom,
-        message: '내용은 최소 10자 이상이어야 합니다.',
+        message: '내용 또는 첨부 파일을 입력해 주세요.',
       });
     }
